Update policeman.properties
Update German local to include the new strings for the new request types. Add the missing "r" to the string I meant to add it to last time. Replace translation for styles to match with the source that says stylesheets.

Please notify if I missed anything that has been added, that is not yet translated.